- Returned skills[skillid][level].x and .y to the equivalent of a short
git-svn-id: svn://70.113.44.214/VanaMirror/trunk@1302 09c1253f-3d09-8e40-9719-4320e56601b5

+++ src/ChannelServer/Skills.h
@@ -40,8 +40,8 @@ struct SkillLevelInfo {
 	int16_t itemcount;
 	int16_t bulletcon;
 	int16_t moneycon;
-	int8_t x;
-	int8_t y;
+	int16_t x;
+	int16_t y;
 	int16_t speed;
 	int16_t jump;
 	int16_t watk;
